Add useHighlight hook along with its documentation and tests as per the guidelines mentioned in the contributing document
